#38 refactoring entities folder
