+++ lib/data/settings/settings_adapter.dart
@@ -1,4 +1,4 @@
-import 'package:dog_sports_diary/domain/entities/settings.dart';
+import 'package:dog_sports_diary/domain/value_objects/settings.dart';
 import 'package:hive/hive.dart';
 import 'dart:convert';
