+++ lib/domain/entities/dog.dart
@@ -1,6 +1,6 @@
 import 'package:dog_sports_diary/domain/abstractions/entity.dart';
-import 'package:dog_sports_diary/domain/entities/sports.dart';
-import 'package:dog_sports_diary/domain/entities/sports_classes.dart';
+import 'package:dog_sports_diary/domain/value_objects/sports.dart';
+import 'package:dog_sports_diary/domain/value_objects/sports_classes.dart';
 
 class Dog implements Entity{
   @override
