+++ lib/domain/value_objects/exercise.dart
