+++ lib/domain/value_objects/ranking.dart
