+++ lib/domain/value_objects/sports_classes.dart
