+++ pubspec.yaml
@@ -42,6 +42,7 @@ dependencies:
   device_info_plus: ^10.1.0
   dropdown_button2: ^3.0.0-beta.16
   file_picker: ^8.0.6
+  fl_chart: ^0.69.0
   flutter_launcher_icons: ^0.13.1
   flutter_rating: ^2.0.2
   flutter_slidable: ^3.1.1
